[AMQCPP-58] - fixed compiler warning on msdev
git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/activemq-cpp/trunk/activemq-cpp@500563 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/network/TcpSocket.cpp
@@ -145,7 +145,7 @@ void TcpSocket::connect(const char* host, int port) throw ( SocketException )
         }
         
         // Create the socket.
-        checkResult( socketHandle = ::socket(AF_INET, SOCK_STREAM, 0) );
+        checkResult( (int)(socketHandle = ::socket(AF_INET, SOCK_STREAM, 0)) );
        
         // Check port value.
         if (port <= 0 || port > 65535) {
